Json Manipulation and UI Work
I was able to get the save_state and load_file methods working.
save_state uses the save_json.py with the GameState class to convert the
state of the manager to a dictionary that can be written to a json file.
Load_file loads a file from the data directory and loads it into the
manager.

I changed the theme to dark mode and changed some colors to make it
more.

I added some widgets to show the current creatures in the manager and
added an input box along with damage and heal buttons. They don't work
yet but the list does actively update along with the add/rm combatant
methods.